Chore: search_service 리팩토링 #49
