+++ program/search_service.py
@@ -1,15 +1,90 @@
 import requests
 import xmltodict
 from django.conf import settings
-from django.http import JsonResponse
-from user.models import User
+from datetime import datetime
+import json
 
-def callByKeyword(keyword, actPlace=None):
+from pathlib import Path
+CURRENT_PATH = Path(__file__).parent.absolute()
+
+def findTagCode(tagName):
+    # split tagname by ' > '
+    tagNameHigh = tagName.split(' > ')[0]
+    tagNameLow = tagName.split(' > ')[1]
+    # find tagCodeHigh from tagList.json
+    with open(CURRENT_PATH / 'tagList.json', 'r', encoding='utf-8') as f:
+        tagList = json.load(f)
+    for tag in tagList:
+        if tag['hignClsNm'] == tagNameHigh and tag['lowClsNm'] == tagNameLow:
+            tagCodeHigh = tag['highClsCd']
+            tagCodeLow = tag['lowClsCd']
+            return tagCodeLow
+    return None
+
+def callByKeyword(keyword, actPlace=None, tagCode=None, areaCode=None):
     url = 'http://openapi.1365.go.kr/openapi/service/rest/VolunteerPartcptnService/getVltrSearchWordList'
+    
+    if type(tagCode) is list and type(areaCode) is list:
+        result = []
+        for tag in tagCode:
+            for area in areaCode:
+                params = {
+                    'keyword' : keyword,
+                    'schCateGu' : 'all',
+                    'actPlace' : actPlace,
+                    'schSign1' : area,
+                    'nanmClCode' : tag,
+                    'numOfRows' : 50
+                    }
+                response = requests.get(url, params=params)
+                parsed_xml = xmltodict.parse(response.text)
+                if parsed_xml["response"]["header"]["resultCode"] == "00":
+                    result = []
+                    itemsList = parsed_xml["response"]["body"]["items"]
+                    if itemsList is None:
+                        return None
+                    items = itemsList['item']
+                    for item in items:
+                        temp = {}
+                        temp['title'] = item.get('progrmSj')
+                        temp['place'] = item.get('actPlace')
+                        temp['progrmRegistNo'] = item.get('progrmRegistNo')
+                        temp['tagCode'] = tagCode
+                        temp['areaCode'] = areaCode
+                        temp['recruitInstitute'] = item.get('nanmmbyNm')
+                        temp['url'] = item.get('url')
+
+                        # yyyymmdd -> yyyy/mm/dd
+                        unparsedRecStartDate = item.get('noticeBgnde')
+                        temp['recruitStart'] = unparsedRecStartDate[0:4] + '/' + unparsedRecStartDate[4:6] + '/' + unparsedRecStartDate[6:8]
+                        unparsedRecEndDate = item.get('noticeEndde')
+                        temp['recruitEnd'] = unparsedRecEndDate[0:4] + '/' + unparsedRecEndDate[4:6] + '/' + unparsedRecEndDate[6:8]
+
+                        # yyyymmdd -> yyyy/mm/dd-hh:mm:ss
+                        unparsedActStartDate = item.get('progrmBgnde')
+                        unparsedActStartTime = item.get('actBeginTm')
+                        # if unparsedActStartTime is smaller than 10, add 0
+                        if int(unparsedActStartTime) < 10:
+                            unparsedActStartTime = '0' + unparsedActStartTime
+                        temp['actStart'] = unparsedActStartDate[0:4] + '/' + unparsedActStartDate[4:6] + '/' + unparsedActStartDate[6:8] + '-' + unparsedActStartTime + ':00:00'
+                        unparsedActEndDate = item.get('progrmEndde')
+                        unparsedActEndTime = item.get('actEndTm')
+                        temp['actEnd'] = unparsedActEndDate[0:4] + '/' + unparsedActEndDate[4:6] + '/'+ unparsedActEndDate[6:8] + '-' + unparsedActEndTime + ':00:00'
+
+                        # d-day 계산 -> recruitEnd - today
+                        temp['dday'] = (datetime.strptime(temp['recruitEnd'], '%Y/%m/%d') - datetime.today()).days
+                        result.append(temp)
+                    return result
+                else:
+                    return None
+        
     params = {
         'keyword' : keyword,
         'schCateGu' : 'all',
-        'actPlace' : actPlace
+        'actPlace' : actPlace,
+        'schSign1' : areaCode,
+        'nanmClCode' : tagCode,
+        'numOfRows' : 50
         }
     response = requests.get(url, params=params)
     parsed_xml = xmltodict.parse(response.text)
@@ -31,15 +106,39 @@ def callByKeyword(keyword, actPlace=None):
             temp['progrmRegistNo'] = items['progrmRegistNo']
 
             result.append(temp)
-        # print(items)
-        else:
-            for item in items:
-                temp = {}
-                temp['title'] = item.get('progrmSj')
-                temp['place'] = item.get('actPlace')
-                temp['progrmRegistNo'] = item.get('progrmRegistNo')
+            return result
 
-                result.append(temp)
+        for item in items:
+            temp = {}
+            temp['title'] = item.get('progrmSj')
+            temp['place'] = item.get('actPlace')
+            temp['progrmRegistNo'] = item.get('progrmRegistNo')
+            temp['tagCode'] = tagCode
+            temp['areaCode'] = areaCode
+            temp['recruitInstitute'] = item.get('nanmmbyNm')
+            temp['url'] = item.get('url')
+
+            # yyyymmdd -> yyyy/mm/dd
+            unparsedRecStartDate = item.get('noticeBgnde')
+            temp['recruitStart'] = unparsedRecStartDate[0:4] + '/' + unparsedRecStartDate[4:6] + '/' + unparsedRecStartDate[6:8]
+            unparsedRecEndDate = item.get('noticeEndde')
+            temp['recruitEnd'] = unparsedRecEndDate[0:4] + '/' + unparsedRecEndDate[4:6] + '/' + unparsedRecEndDate[6:8]
+
+            # yyyymmdd -> yyyy/mm/dd-hh:mm:ss
+            unparsedActStartDate = item.get('progrmBgnde')
+            unparsedActStartTime = item.get('actBeginTm')
+            # if unparsedActStartTime is smaller than 10, add 0
+            if int(unparsedActStartTime) < 10:
+                unparsedActStartTime = '0' + unparsedActStartTime
+            temp['actStart'] = unparsedActStartDate[0:4] + '/' + unparsedActStartDate[4:6] + '/' + unparsedActStartDate[6:8] + '-' + unparsedActStartTime + ':00:00'
+            unparsedActEndDate = item.get('progrmEndde')
+            unparsedActEndTime = item.get('actEndTm')
+            temp['actEnd'] = unparsedActEndDate[0:4] + '/' + unparsedActEndDate[4:6] + '/'+ unparsedActEndDate[6:8] + '-' + unparsedActEndTime + ':00:00'
+
+            # d-day 계산 -> recruitEnd - today
+            temp['dday'] = (datetime.strptime(temp['recruitEnd'], '%Y/%m/%d') - datetime.today()).days
+
+            result.append(temp)
         # print(result)
         return result
     else:
@@ -48,9 +147,10 @@ def callByKeyword(keyword, actPlace=None):
 def callByArea(keyword):
     url = 'http://openapi.1365.go.kr/openapi/service/rest/VolunteerPartcptnService/getVltrAreaList'
     params = {
-    'serviceKey' : settings.VOL_API_KEY,
-    'schCateGu' : 'all',
-    'schSign1' : keyword, # 지역코드 (구군)    
+        'serviceKey' : settings.VOL_API_KEY,
+        'schCateGu' : 'all',
+        'schSign1' : keyword, # 지역코드 (구군),
+        'numOfRows' : 50    
     }
     # 시도 단위로 검색하려면, schSign1 대신
     # 'schSido' : keyword # 지역코드(시도)
@@ -91,7 +191,7 @@ def callByRegistNo(registNo):
     url = 'http://openapi.1365.go.kr/openapi/service/rest/VolunteerPartcptnService/getVltrPartcptnItem'
     params ={
         'serviceKey' : settings.VOL_API_KEY,
-        'progrmRegistNo' : registNo
+        'progrmRegistNo' : registNo,
     }
     response = requests.get(url, params=params)
     parsed_xml = xmltodict.parse(response.text)
@@ -133,11 +233,14 @@ def callByRegistNo(registNo):
         temp['recruitStart'] = recruitStart
         temp['recruitEnd'] = recruitEnd
         temp['recruitInstitute'] = item.get('mnnstNm')
-        temp['registerInstitute'] = item.get('nanmmbyNm')
+        temp['registerInstiute'] = item.get('nanmmbyNm')
         temp['maxPerson'] = item.get('rcritNmpr')
         temp['content']= item.get('progrmCn')
         temp['progrmRegistNo'] = registNo
-        temp['tagName']= item.get('srvcClCode')
+        temp['tagName'] = item.get('srvcClCode')
+        temp['url'] = f"https://www.1365.go.kr/vols/P9210/partcptn/timeCptn.do?type=show&progrmRegistNo={registNo}"
+        temp['areaCode'] = item.get('gugunCd')
+        temp['tagCode'] = findTagCode(temp['tagName'])
 
         return temp
     else:
