+++ program/views.py
@@ -9,6 +9,7 @@ from rest_framework import status
 from rest_framework.response import Response
 
 from .service import *
+from .search_service import *
 from .response import *
 from .serializers import *
 # from .search_service import *
